Get rid of xmlns:ns and ns:-prefixed attributes
It seems that `xml-conduit` treats attributes with namespaces
differently than `xml` did and produces the following output:

```xml
<content xmlns:ns="http://www.w3.org/2005/Atom" ns:type="html">
```
instead of

```xml
<content type="html">
```

which seems to confuse some feed readers, notably
<https://github.com/lwindolf/liferea>.

If I'm interpreting https://stackoverflow.com/a/46865/3407728 correctly,
it seems that these attributes indeed aren't meant to be namespaced.
Indeed, if I inspect some `<a href="whatever">` element in a browser
using the javascript console, the element does have a namespace but the
attribute does not.

Therefore, drop the namespace from attributes.